requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --index-url=https://artefact-repository/repository/pypi/simple --resolver=backtracking requirements.in
#
--index-url https://artefact-repository/repository/pypi/simple

alembic==1.12.0
    # via
    #   -r requirements.in
    #   flask-migrate
blinker==1.6.2
    # via flask
certifi==2023.7.22
    # via requests
charset-normalizer==3.2.0
    # via requests
click==8.1.7
    # via flask
flask==2.3.3
    # via
    #   -r requirements.in
    #   flask-logconfig
    #   flask-migrate
    #   flask-script
    #   flask-sqlalchemy
flask-logconfig==0.4.2
    # via -r requirements.in
flask-migrate==4.0.4
    # via -r requirements.in
flask-script==2.0.6
    # via -r requirements.in
flask-sqlalchemy==3.0.5
    # via
    #   -r requirements.in
    #   flask-migrate
geoalchemy2==0.14.1
    # via -r requirements.in
greenlet==2.0.2
    # via sqlalchemy
gunicorn==21.2.0
    # via -r requirements.in
idna==3.4
    # via requests
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
jwt-validation @ git+http://internal-git-host/llc-beta/jwt-validation.git@v1.2.3
    # via -r requirements.in
logconfig==0.4.0
    # via flask-logconfig
logutils==0.3.5
    # via logconfig
mako==1.2.4
    # via alembic
markupsafe==2.1.3
    # via
    #   jinja2
    #   mako
    #   werkzeug
marshmallow==3.20.1
    # via jwt-validation
numpy==1.25.2
    # via shapely
packaging==23.1
    # via
    #   geoalchemy2
    #   gunicorn
    #   marshmallow
psycopg2==2.9.7
    # via -r requirements.in
python-dateutil==2.8.2
    # via -r requirements.in
pyyaml==6.0.1
    # via logconfig
requests==2.31.0
    # via
    #   -r requirements.in
    #   jwt-validation
shapely==2.0.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sqlalchemy==1.4.46
    # via
    #   -r requirements.in
    #   alembic
    #   flask-sqlalchemy
    #   geoalchemy2
typing-extensions==4.7.1
    # via alembic
urllib3==2.0.4
    # via requests
werkzeug==2.3.7
    # via
    #   -r requirements.in
    #   flask
